requirements.txt
# Core Data Science Libraries
pandas==2.0.3
numpy==1.24.3
scipy==1.11.1
scikit-learn==1.3.0
statsmodels==0.14.0

# Database Connectivity
psycopg2-binary==2.9.7
pymongo==4.4.1
sqlalchemy==2.0.19

# Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0
streamlit==1.25.0
bokeh==3.2.2

# ETL and Data Processing
apache-airflow==2.6.3
great-expectations==0.17.7
dask==2023.7.0

# Jupyter and Development
jupyter==1.0.0
jupyterlab==4.0.7
ipykernel==6.25.0

# Testing
pytest==7.4.0
pytest-cov==4.1.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
tqdm==4.65.0
faker==19.3.1

# API and Web
fastapi==0.101.1
uvicorn==0.23.2
requests==2.31.0

# Data Validation
cerberus==1.3.4
marshmallow==3.20.1
